+++ webknossos/webknossos/client/api_client/datastore_api_client.py
@@ -27,8 +27,9 @@ class DatastoreApiClient(AbstractApiClient):
         datastore_base_url: str,
         timeout_seconds: float,
         headers: dict[str, str] | None = None,
+        webknossos_api_version: int = 15,
     ):
-        super().__init__(timeout_seconds, headers)
+        super().__init__(timeout_seconds, headers, webknossos_api_version)
         self.datastore_base_url = datastore_base_url.rstrip("/")
 
     @property