+++ webknossos/webknossos/client/api_client/wk_api_client.py
@@ -60,8 +60,9 @@ class WkApiClient(AbstractApiClient):
         base_wk_url: str,
         timeout_seconds: float,
         headers: dict[str, str] | None = None,
+        webknossos_api_version: int = 15,
     ):
-        super().__init__(timeout_seconds, headers)
+        super().__init__(timeout_seconds, headers, webknossos_api_version)
         self.base_wk_url = base_wk_url.rstrip("/")
 
     @property